tests/wpt/tests/dom/nodes/getElementsByClassName-14.htm

<!-- quirks mode -->
<html class="a A">
 <head>
  <title>document.getElementsByClassName(): case-insensitive (quirks mode)</title>
  <link rel="help" href="https://dom.spec.whatwg.org/#concept-getelementsbyclassname">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body class="a a">
  <div id="log"></div>
  <div class="k"></div>
  <div class="K"></div>
  <div class="&#x212a;" id="kelvin"></div>
  <script>
test(function() {
  assert_array_equals(document.getElementsByClassName("A a"),
                      [document.documentElement, document.body]);
})

test(function() {
  assert_array_equals(document.getElementsByClassName("\u212a"),
                      [document.getElementById("kelvin")]);
}, 'Unicode-case should be sensitive even in quirks mode.');
  </script>
 </body>
</html>
